Feature/interests get#22

+++ src/main/java/com/monew/monew_server/domain/interest/dto/InterestDto.java
@@ -7,6 +7,6 @@ public record InterestDto(
     UUID id,
     String name,
     List<String> keywords,
-    Integer subscriberCount,
+    Long subscriberCount,
     Boolean subscribedByMe
 ) {}
